Refactor client functionality into main.py; remove client.py as it is no longer needed.
